Improve Depolyment interface (#82)
* Support partial use of builtin rules

Bazel allows a gradual move from builtin rules to starlark rules.
During that process, it is possible to have projects where only
some rules are builtin (controllable by appropriate flags in the rc
file) wheras as others need to be imported. To adequately support
this situation, change the deployment configuration to accept a
set of languages for which (still) to use the builtin rules.

* Add transformation from known names for languages to repo mapping

Some rules changed their name over time, so it is not obvious for
a user of the deployment library which name is used in our aspect
code. To still allow a stable way of constructing the appropriate
repository mapping add a function generating it from a mapping of
our language enum to the repository names to be used.

* Deploy: only touch a file if it does not have the correct content already

Touching a bzl file invalidates large parts of the analysis cache,
especially if it defines an aspect to be run over all targets.
Therefore, only touch files if the content has to be changed.

* Rename Langauges to Rules to better reflect its usage

... and also make the rule-set name a property of the enum.

* Aspect config: request renaming as map from rule

... rather than from strings. In this way, requesting a depolyment
does not require the implementation detail of which rule-set name
is used in case of rules that changed their name.

+++ sdk/lib/deploy.kt
@@ -22,7 +22,9 @@ import java.nio.file.Files
 import java.nio.file.Path
 import java.nio.file.StandardOpenOption
 import java.util.zip.ZipInputStream
+import kotlin.io.path.exists
 import kotlin.io.path.extension
+import kotlin.io.path.readText
 
 /**
  * Location of the bundled aspect archive inside the jar.
@@ -37,11 +39,11 @@ data class AspectConfig(
   /**
    * A mapping from default repo names to a specific replacement e.g., conical repo name.
    */
-  val repoMapping: Map<String, String>,
+  val repoMapping: Map<Rules, String>,
   /**
-   * Whether to use builtin rules i.e. whether to strip rule set loads.
+   * Languages for which to use the builtin rule, i.e., for which to strip rule set loads.
    */
-  val useBuiltin: Boolean,
+  val useBuiltin: Set<Rules>,
 )
 
 /**
@@ -71,10 +73,10 @@ fun deployAspectZip(
     TransformExternalRepositories(config.repoMapping),
   )
 
-  if (config.useBuiltin) {
+  if (Rules.CC in config.useBuiltin) {
     transformers.add(TransformCcToolchainType)
-    transformers.add(TransformBuiltinRules)
   }
+  transformers.add(TransformBuiltinRules(config.useBuiltin))
 
   val archiveStream = if (archiveZip == null) {
     config.javaClass.getResourceAsStream(BUNDLED_ASPECT)
@@ -102,13 +104,16 @@ private fun extractZipArchive(
       if (entry.isDirectory) {
         Files.createDirectories(target)
       } else {
-        Files.writeString(
-          target,
-          transformFile(stream.nonClosing(), transformers),
-          Charsets.UTF_8,
-          StandardOpenOption.CREATE,
-          StandardOpenOption.TRUNCATE_EXISTING,
-        )
+        val payload = transformFile(stream.nonClosing(), transformers)
+        if (!target.exists() || target.readText() != payload) {
+          Files.writeString(
+            target,
+            payload,
+            Charsets.UTF_8,
+            StandardOpenOption.CREATE,
+            StandardOpenOption.TRUNCATE_EXISTING,
+          )
+        }
       }
     }
   }
